+++ tests/testthat/test-select-ranks.R
@@ -13,23 +13,25 @@ test_that("Default execution of nmarank without condition", {
 
 A = condition("retainOrder", c("Placebo", "Salmeterol", "SFC"))
 B = condition("specificPosition", "Placebo", 1)
-C = condition("sameHierarchy", c("Placebo", "Fluticasone", "Salmeterol", "SFC"))
+C = condition("sameHierarchy",
+              c("Placebo", "Fluticasone", "Salmeterol", "SFC"))
 D = condition("betterEqual", "Fluticasone", 2)
 G = condition("betterEqual", "Placebo", 2)
 
 test_that("Build selection tree", {
   st = (A %AND% (B %OR% (C %XOR% D)))
-  expect_equal( st$root$isRoot
-               , T
-               )
+  expect_equal(st$root$isRoot, TRUE)
 })
 
-test_that("small.values='good' should give Placebo last for this mortality outcome", {
+test_that(paste("small.values = 'desirable' should give",
+                "Placebo last for this mortality outcome"), {
   st1 <- condition("specificPosition", "Placebo", 1)
-  placeboFirst = nmarank(net1, condition=st1,small.values="good")$probabilityOfSelection
+  placeboFirst = nmarank(net1, condition = st1,
+                         small.values = "desirable")$probabilityOfSelection
   st2 <- condition("specificPosition", "Placebo", 4)
-  placeboLast = nmarank(net1, condition=st2,small.values="good")$probabilityOfSelection
-  expect_true(placeboLast>placeboFirst)
+  placeboLast = nmarank(net1, condition = st2,
+                        small.values = "desirable")$probabilityOfSelection
+  expect_true(placeboLast > placeboFirst)
 })
 
 test_that("check Selection tree", {
@@ -37,10 +39,10 @@ test_that("check Selection tree", {
   st1 = (B %XOR% (C %OR% (D %AND% G))) %OR% A
   effs <- nmarank:::nmaEffects(net1$TE.random, net1$Cov.random)
   ranksrow = effs$TE
-  holds = selectionHolds(st, small.values="bad", ranksrow)
+  holds = selectionHolds(st, small.values = "undesirable", ranksrow)
   expect_true(holds)
-  expect_equal(selectionHolds(st, small.values="bad", ranksrow),
-               selectionHolds(st1, small.values="bad", ranksrow))
+  expect_equal(selectionHolds(st, small.values = "undesirable", ranksrow),
+               selectionHolds(st1, small.values = "undesirable", ranksrow))
 })
 
 test_that("Commutative selections", {